Feat: update syncing process

+++ src/repositories/block.repository.js
@@ -12,6 +12,7 @@ import { summaryRepository } from './summary.repository'
 import { addressRepository } from './address.repository'
 import { isAddress } from '@rsksmart/rsk-utils/dist/addresses'
 import { balancesRepository } from './balances.repository'
+import { statusRepository } from './status.repository'
 
 export const blockRepository = {
   async findOne (query = {}, project = {}, collection) {
@@ -40,7 +41,7 @@ export const blockRepository = {
     return transactionQueries
   },
   async saveBlockData (data) {
-    const { block, transactions, internalTransactions, events, tokenAddresses, addresses, balances } = data
+    const { block, transactions, internalTransactions, events, tokenAddresses, addresses, balances, status } = data
     const transactionQueries = []
 
     // insert block
@@ -59,8 +60,12 @@ export const blockRepository = {
     transactionQueries.push(...balancesRepository.insertMany(balances))
 
     // insert txs and delete pendings
-    for (const tx of transactions) {
-      transactionQueries.push(...txRepository.insertOne(tx), ...txPendingRepository.deleteOne({ hash: tx.hash }))
+    if (!transactions.length && block.number > 0) {
+      throw new Error(`Couldn't get transactions for block ${block.number}`)
+    } else {
+      for (const tx of transactions) {
+        transactionQueries.push(...txRepository.insertOne(tx), ...txPendingRepository.deleteOne({ hash: tx.hash }))
+      }
     }
 
     // insert internal transactions
@@ -83,6 +88,11 @@ export const blockRepository = {
 
     // save block summary
     transactionQueries.push(...summaryRepository.insertOne(data))
+
+    // insert status
+    if (status) {
+      transactionQueries.push(...statusRepository.insertOne(status))
+    }
 
     const res = prismaClient.$transaction(transactionQueries)
